create ProjectOverview component
